+++ packages/storybook-html/stories/components/breadcrumb/breadcrumb.stories.js
@@ -14,7 +14,7 @@ export default {
 export const Breadcrumb = {
   render: () => `
 <div class="uq-breadcrumbs">
-  <div class="uq-container">
+  <div class="uq-container uq-container--flex">
     <nav class="uq-breadcrumb" aria-label="Breadcrumb">
       <ol class="uq-breadcrumb__list">
         <li class="uq-breadcrumb__item">