+++ packages/storybook-html/stories/components/header/header.stories.js
@@ -32,45 +32,41 @@ export const HeaderInterim = {
     const headerElem = canvasElement.querySelector(".uq-header");
     if (headerElem) new header(headerElem);
 
-    document.addEventListener("DOMContentLoaded", function () {
-      const menuLeftElem = document.getElementById("global-mobile-nav");
+    const menuLeftElem = headerElem.querySelector("#global-mobile-nav");
+    const menuLeft = new SlideMenu(menuLeftElem, {
+      position: "left",
+      submenuLinkAfter: " ",
+      backLinkBefore: " ",
+    });
 
-      const menuLeft = new SlideMenu(menuLeftElem, {
-        position: "left",
-        submenuLinkAfter: " ",
-        backLinkBefore: " ",
+    const searchToggle = headerElem.querySelector(
+      ".nav-primary__search-toggle",
+    );
+    searchToggle.addEventListener("click", () => {
+      menuLeft.close();
+    });
+
+    var slideMenuBackButtons = headerElem.querySelectorAll(
+      ".slide-menu__backlink, .global-mobile-nav__audience-link",
+    );
+
+    Array.prototype.forEach.call(slideMenuBackButtons, function (el, i) {
+      el.addEventListener("click", () => {
+        menuLeftElem.scrollTop = 0;
       });
+    });
 
-      this.searchToggle = document.querySelector(".nav-primary__search-toggle");
-
-      this.searchToggle.addEventListener("click", () => {
-        menuLeft.close();
-      });
-
-      var slideMenuBackButtons = document.querySelectorAll(
-        ".slide-menu__backlink, .global-mobile-nav__audience-link"
-      );
-
-      Array.prototype.forEach.call(slideMenuBackButtons, function (el, i) {
-        el.addEventListener("click", () => {
-          document.querySelector(".global-mobile-nav").scrollTop = 0;
-        });
-      });
-
-      // Responsive Resize Close menu and update toggles
-      window.addEventListener("resize", (event) => {
-        // Target Resize of LG ($screen-lg, 64rem, 1024px).
-        if (window.innerWidth > 1024) {
-          menuLeft.close(true);
-          //reset the menu toggle after closing.
-          this.mainNavToggle = document.querySelector(".nav-primary__toggle");
-          this.mainNavToggle.classList.remove(
-            "nav-primary__menu-toggle--is-open"
-          );
-          this.body = document.querySelector("body");
-          this.body.classList.remove("no-scroll");
-        }
-      });
+    // Responsive Resize Close menu and update toggles
+    window.addEventListener("resize", (event) => {
+      // Target Resize of LG ($screen-lg, 64rem, 1024px).
+      if (window.innerWidth > 1024) {
+        menuLeft.close(true);
+        //reset the menu toggle after closing.
+        const mainNavToggle = headerElem.querySelector(".nav-primary__toggle");
+        mainNavToggle.classList.remove("nav-primary__menu-toggle--is-open");
+        const body = document.querySelector("body");
+        body.classList.remove("no-scroll");
+      }
     });
   },
 
@@ -85,45 +81,41 @@ export const Header = {
     const headerElem = canvasElement.querySelector(".uq-header");
     if (headerElem) new header(headerElem);
 
-    document.addEventListener("DOMContentLoaded", function () {
-      const menuLeftElem = document.getElementById("global-mobile-nav");
+    const menuLeftElem = headerElem.querySelector("#global-mobile-nav");
+    const menuLeft = new SlideMenu(menuLeftElem, {
+      position: "left",
+      submenuLinkAfter: " ",
+      backLinkBefore: " ",
+    });
 
-      const menuLeft = new SlideMenu(menuLeftElem, {
-        position: "left",
-        submenuLinkAfter: " ",
-        backLinkBefore: " ",
+    const searchToggle = headerElem.querySelector(
+      ".nav-primary__search-toggle",
+    );
+    searchToggle.addEventListener("click", () => {
+      menuLeft.close();
+    });
+
+    var slideMenuBackButtons = headerElem.querySelectorAll(
+      ".slide-menu__backlink, .global-mobile-nav__audience-link",
+    );
+
+    Array.prototype.forEach.call(slideMenuBackButtons, function (el, i) {
+      el.addEventListener("click", () => {
+        menuLeftElem.scrollTop = 0;
       });
+    });
 
-      this.searchToggle = document.querySelector(".nav-primary__search-toggle");
-
-      this.searchToggle.addEventListener("click", () => {
-        menuLeft.close();
-      });
-
-      var slideMenuBackButtons = document.querySelectorAll(
-        ".slide-menu__backlink, .global-mobile-nav__audience-link"
-      );
-
-      Array.prototype.forEach.call(slideMenuBackButtons, function (el, i) {
-        el.addEventListener("click", () => {
-          document.querySelector(".global-mobile-nav").scrollTop = 0;
-        });
-      });
-
-      // Responsive Resize Close menu and update toggles
-      window.addEventListener("resize", (event) => {
-        // Target Resize of LG ($screen-lg, 64rem, 1024px).
-        if (window.innerWidth > 1024) {
-          menuLeft.close(true);
-          //reset the menu toggle after closing.
-          this.mainNavToggle = document.querySelector(".nav-primary__toggle");
-          this.mainNavToggle.classList.remove(
-            "nav-primary__menu-toggle--is-open"
-          );
-          this.body = document.querySelector("body");
-          this.body.classList.remove("no-scroll");
-        }
-      });
+    // Responsive Resize Close menu and update toggles
+    window.addEventListener("resize", (event) => {
+      // Target Resize of LG ($screen-lg, 64rem, 1024px).
+      if (window.innerWidth > 1024) {
+        menuLeft.close(true);
+        //reset the menu toggle after closing.
+        const mainNavToggle = headerElem.querySelector(".nav-primary__toggle");
+        mainNavToggle.classList.remove("nav-primary__menu-toggle--is-open");
+        const body = document.querySelector("body");
+        body.classList.remove("no-scroll");
+      }
     });
   },
 
@@ -158,7 +150,7 @@ export const searchGlobal = {
               <label for="edit-q" class="visually-hidden uq-header__search-query-label">Search term</label>
               <input type="text" id="edit-q" name="q" value="" maxlength="128" placeholder="Search by keyword" class="uq-header__search-query-input" data-gtm-trigger="change" data-gtm-form-search="">
               <span class="uq-header__search-query-button">
-                <input type="submit" name="op" value="Search" class="uq-header__search-query-submit" data-gtm-trigger="click">
+                <input type="submit" value="Search" class="uq-header__search-query-submit" data-gtm-trigger="click">
               </span>
             </div>
           </fieldset>
@@ -186,7 +178,7 @@ export const searchLocal = {
           <label for="edit-q" class="visually-hidden uq-header__search-query-label">Search term</label>
           <input type="text" id="edit-q" name="q" value="" maxlength="128" placeholder="Search by keyword" class="uq-header__search-query-input" data-gtm-trigger="change" data-gtm-form-search="">
           <span class="uq-header__search-query-button">
-            <input type="submit" name="op" value="Search" class="uq-header__search-query-submit" data-gtm-trigger="click">
+            <input type="submit" value="Search" class="uq-header__search-query-submit" data-gtm-trigger="click">
           </span>
         </div>
         <div class="uq-header__search-range">