+++ packages/storybook-html/stories/components/hero-basic/hero-basic.stories.js
@@ -88,7 +88,7 @@ export const HeroBasic = {
               "uq-tabs__tab--active": index === 0,
             })}">
               ${name}
-            </a>`
+            </a>`,
               )
               .join("")}
           </div>