+++ packages/footer-legacy/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4-alpha.0](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.3...@uqds/footer-legacy@0.1.4-alpha.0) (2024-09-04)
+
+**Note:** Version bump only for package @uqds/footer-legacy
+
+## [0.1.3](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.2-alpha.0...@uqds/footer-legacy@0.1.3) (2024-09-04)
+
+**Note:** Version bump only for package @uqds/footer-legacy
+
+## [0.1.2](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.2-alpha.0...@uqds/footer-legacy@0.1.2) (2024-09-04)
+
+**Note:** Version bump only for package @uqds/footer-legacy
+
+## [0.1.2-alpha.0](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.0-alpha.1...@uqds/footer-legacy@0.1.2-alpha.0) (2024-05-03)
+
+**Note:** Version bump only for package @uqds/footer-legacy
+
+## [0.1.1](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.0-alpha.1...@uqds/footer-legacy@0.1.1) (2024-05-03)
+
+**Note:** Version bump only for package @uqds/footer-legacy
+
+# [0.1.0](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.0-alpha.1...@uqds/footer-legacy@0.1.0) (2024-05-03)
+
+**Note:** Version bump only for package @uqds/footer-legacy
+
 # [0.1.0-alpha.1](https://github.com/uq-its-ss/design-system/compare/@uqds/footer-legacy@0.1.0-alpha.0...@uqds/footer-legacy@0.1.0-alpha.1) (2024-04-05)
 
 **Note:** Version bump only for package @uqds/footer-legacy